requirements.txt
Flask>=3.1,<4
Flask-SQLAlchemy>=3.1,<4
Flask-WTF>=1.2,<2
SQLAlchemy>=2.0,<2.1
WTForms>=3.2,<4
gunicorn>=25,<26
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
requests>=2.32,<3
